+++ Sandbox/GRCookbook/GRCookbookPlugins.h
@@ -112,28 +112,3 @@ private:
 
     Zelo::Core::ECS::Entity *entity{};
 };
-
-namespace FinalScene {
-class Ch10FinalPlugin : public Plugin {
-public:
-
-public:
-    const std::string &getName() const override;;
-
-    void install() override;
-
-    void uninstall() override;
-
-    void initialize() override;
-
-    void update() override;
-
-    void render() override;
-
-private:
-    struct Impl;
-    std::shared_ptr<Impl> pimpl{};
-
-    Zelo::Core::ECS::Entity *entity{};
-};
-}
